Improve initial load of flow editor
`FlowPart.Edit` will build an editor for each widget that can be added
to a flow in order to get any resources (Stylesheets/JavaScript) loaded
on the page for when the widget is added. This improvement changes the
`BuildEditor` endpoint response and JavaScript for handling the response
to handle loading the Stylesheets/JavaScript files when they haven't
been added to the page already.